Implemented receipt, purpose & payment entity TODO   fix 'Unable to guess FieldBridge for importDate'   test Receipt entity & Payment

+++ project/project-et-ejb/src/main/java/cz/muni/fi/pv243/et/data/impl/PaymentRepositoryImpl.java
@@ -3,29 +3,49 @@ package cz.muni.fi.pv243.et.data.impl;
 import cz.muni.fi.pv243.et.data.PaymentRepository;
 import cz.muni.fi.pv243.et.model.Payment;
 
+import javax.ejb.Stateless;
 import javax.inject.Inject;
 import javax.persistence.EntityManager;
 
 
+@Stateless
 public class PaymentRepositoryImpl implements PaymentRepository {
 
     @Inject
     private EntityManager em;
 
     @Override
-    public void createPayment(Payment payment) {
+    public void create(Payment payment) {
+        if (payment == null) {
+            throw new IllegalArgumentException("payment is null");
+        }
+        if (payment.getId() != null) {
+            throw new IllegalArgumentException("payment.id is not null");
+        }
         em.persist(payment);
-        em.flush();
+
     }
 
     @Override
-    public void updatePayment(Payment payment) {
+    public void update(Payment payment) {
+        if (payment == null) {
+            throw new IllegalArgumentException("payment is null");
+        }
+        if (payment.getId() == null) {
+            throw new IllegalArgumentException("payment.id is null - not persisted");
+        }
         em.merge(payment);
         em.flush();
     }
 
     @Override
-    public void removePayment(Payment payment) {
+    public void remove(Payment payment) {
+        if (payment == null) {
+            throw new IllegalArgumentException("payment is null");
+        }
+        if (payment.getId() == null) {
+            throw new IllegalArgumentException("payment.id is null - not persisted");
+        }
         //em.remove(em.find(Payment.class, payment.getId()));
         em.remove(payment);
         em.flush();
